requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
mouse==0.7.1
openai==0.28.1
python-dotenv==1.0.1 